+++ packages/isaac-typescript-definitions/src/types/classes/ItemPool.d.ts
@@ -38,10 +38,19 @@ declare global {
     GetPill(seed: Seed): PillColor;
 
     /**
+     * Will return the pill effect that corresponds to the passed pill color. This will work
+     * properly even if the player has not yet identified the pill color (by using one or more pills
+     * of that color). It is recommended to always pass the corresponding player because if a player
+     * has Lucky Foot, PHD, Virgo, or False PHD, the resolved pill effect will change from what was
+     * assigned by default at the beginning of the run.
+     *
+     * Returns -1 if passed `PillColor.NULL` (0) or a value of 2048. Returns `PillEffect.BAD_GAS`
+     * (0) if passed an invalid pill color (e.g. 15 through 2047 or 2063+).
+     *
      * @param pillColor
      * @param player Default is undefined.
      */
-    GetPillEffect(pillColor: PillColor, player?: EntityPlayer): PillEffect;
+    GetPillEffect(pillColor: PillColor, player?: EntityPlayer): PillEffect | -1;
 
     /**
      * Note that this function will return `ItemPoolType.NULL` for `RoomType.DEFAULT`, which may be